[Support4] Fix Color-related breakage.
Interestingly, the Color changes found a bug. This isn't valid:

	new ColorDrawable(Resource.Drawable.background)

It's not valid because ColorDrawable wants a color.
Resource.Drawable.background isn't a color; it's a color reference.
(Compare the value of Resource.Drawable.background -- 0x7f020007 -- to
the desired value -- #ff0000.)

To get the actual corresponding color you need to use
Resources.GetColor(int).

+++ Support4/Support4/Activities/FragmentLayoutSupport.cs
@@ -77,7 +77,7 @@ namespace Support4
 				if(dualPane)
 				{
 					// In dual-pane mode, the list view highlights the selected item.
-					ListView.ChoiceMode = (int) ChoiceMode.Single;
+					ListView.ChoiceMode = ChoiceMode.Single;
                 
 	                // Make sure our UI is in the correct state.
 	                ShowDetails(curCheckPosition);
